requirements.txt
pandas
dash
plotly
dash-bootstrap-components
wordcloud
us
gmaps
pillow